On the path to lazy zero-alloc. This removes many allocations at the bufrw layer.

It also gets us a few hundred ops/sec faster for the eager path:

Before
```
1:          PING,  1000/0 min/max/avg/p95:    9/ 352/ 154.68/ 265.55   3211ms total,  6228.59 ops/sec
1:          PING,  1000/0 min/max/avg/p95:   17/ 295/ 146.34/ 253.00   3001ms total,  6664.45 ops/sec
1:          PING,  1000/0 min/max/avg/p95:   21/ 287/ 141.63/ 241.00   2903ms total,  6889.42 ops/sec
1:          PING,  1000/0 min/max/avg/p95:   21/ 282/ 141.81/ 241.00   2908ms total,  6877.58 ops/sec
```

After
```
1:          PING,  1000/0 min/max/avg/p95:    9/ 322/ 143.92/ 246.55   2989ms total,  6691.20 ops/sec
1:          PING,  1000/0 min/max/avg/p95:   10/ 288/ 139.48/ 235.55   2852ms total,  7012.62 ops/sec
1:          PING,  1000/0 min/max/avg/p95:   11/ 278/ 137.48/ 235.55   2812ms total,  7112.38 ops/sec
1:          PING,  1000/0 min/max/avg/p95:   10/ 282/ 141.41/ 242.00   2890ms total,  6920.42 ops/sec
```

r @jcorbin @kriskowal